### 2019-07-08 #21 @whedon
[PRE REVIEW]: Interoperating Deep Learning models with ONNX.jl
**Submitting author:** @ayush-1506 (<a href="http://orcid.org/0000-0003-3550-691X">Ayush Shridhar</a>)
**Repository:** <a href="https://github.com/FluxML/ONNX.jl" target ="_blank">https://github.com/FluxML/ONNX.jl</a>

Thanks for submitting your paper to JuliaCon @ayush-1506. The JuliaCon editor @vchuravy, will work with you on this issue to find a reviewer for your submission before creating the main review issue.

@ayush-1506 if you have any suggestions for potential reviewers then please mention them here in this thread. In addition, [this list of people](http://bit.ly/2V9GSFq) have already agreed to review for JuliaCon and may be suitable for this submission.

The JuliaCon submission bot @whedon is here to help you find and assign reviewers and start the main review. To find out what @whedon can do for you type:
