### 2019-09-23 #21 @vchuravy
Thank you. As soon as Ayush has fixed the paper I will start the review process.

Take a look at https://proceedings.juliacon.org/guide/reviewers/ in the meanwhile and I am happy to answer questions.
