### 2019-09-30 #21 @ayush-1506
@vchuravy Thanks a lot. I'll be targeting the special edition, not the JuliaCon proceedings, since I wasn't able to make it to JuliaCon this year (I created this PR before knowing that I wasn't going to make it). Should I keep this PR open, or move it to another repository for the special edition?
