### 2019-09-30 #21 @vchuravy
@ayush-1506 since you were accepted to JuliaCon, but couldn't make it for unrelated reasons it is up to you if you want to submit this to the proceedings or the special edition. In either case the submission process is the same and I would encourage you to finish it up as soon as possible, since we already have reviewers. Both proceedings and special edition are rolling submission.
