### 2019-11-04 83eed31
modifications to the paper
