+++ paper/paper.tex
@@ -5,27 +5,63 @@
 
 \begin{document}
 
-\input{header}
+\title{Interoperating Deep Learning models with ONNX.jl}
+
+\author[1]{Ayush Shridhar}
+\author[2]{Mike Innes}
+
+\affil[1]{International Institute of Information Technology, Bhubaneswar, India}
+\affil[2]{Julia Computing}
+
+\keywords{Julia, Machine Learning, Deep Learning, Transfer Learning}
 
 \maketitle
 
 \begin{abstract}
 
-Flux\cite{innes:2018} is a machine learning framework, written using the numerical computing language Julia\cite{bezanson2017julia}. The framework makes writing layers as simple as writing mathematical formulae, and it's advanced AD, Zygote\cite{DBLP:journals/corr/abs-1810-07951}, applies automatic differentiation (AD) to seamlessly calculate derivatives and train the model. It makes heavy use of Julia's language and compiler features to carry out code analysis and make optimisations.  For example, Julia's GPU compilation support\cite{besard:2017} can be used to JIT-compile custom GPU kernels for model layers\cite{CuArrays}. Flux also supports a number of a hardware options, from CPUs, GPUs and even including TPUs.\newline
+Flux\cite{innes:2018} is a machine learning framework, written using the numerical computing
+language Julia\cite{bezanson2017julia}. The framework makes writing layers as simple as
+writing mathematical formulae, and it's advanced AD,
+Zygote\cite{DBLP:journals/corr/abs-1810-07951}, applies automatic differentiation (AD) to
+calculate derivatives and train the model. It makes heavy use of Julia's language and
+compiler features to carry out code analysis and make optimisations.  For example, Julia's
+GPU compilation support\cite{besard:2017} can be used to JIT-compile custom GPU kernels for
+model layers\cite{CuArrays}. Flux also supports a number of a hardware options, from CPUs,
+GPUs and even including TPUs.\newline
 
-ONNX.jl is an Open Neural Network Exchange backend for the Flux.jl deep learning framework. ONNX.jl supports directly importing high quality ONNX standard models into Flux, thus saving time and reducing the need for additional computation resources.
+ONNX.jl is an Open Neural Network Exchange backend for the Flux.jl deep learning framework.
+ONNX.jl supports directly importing high quality ONNX standard models into Flux, thus saving
+time and reducing the need for additional computation resources. In this paper we'll look
+into finer details of how ONNX.jl produces Julia code of a model from a serialized file.
 
 \end{abstract}
 
 \section{Introduction}
 
-Open Neural Network Exchange (ONNX) is an open ecosystem that empowers AI developers to choose the right tools as their project evolves. ONNX provides an open source format for AI models, both deep learning and traditional ML. It defines an extensible computation graph model, as well as definitions of built-in operators and standard data types. It provides a set of specifications to convert a model to a basic ONNX format, and another set of specifications to get the model back from this ONNX form.\newline  At a high level, ONNX is designed to allow framework inter-operability. There are many excellent machine learning libraries in various languages — PyTorch, TensorFlow, MXNet, and Caffe are just a few that have become very popular in recent years, but there are many others as well.
+Open Neural Network Exchange (ONNX) is an open ecosystem that empowers AI developers to
+choose the right tools as their project evolves. ONNX provides an open source format for AI
+models, both deep learning and traditional machine learning. It defines an extensible
+computation graph model, as well as definitions of built-in operators and standard data
+types. It provides a set of specifications to convert a model to a basic ONNX format, and
+another set of specifications to get the model back from this ONNX form.\newline  At a high
+level, ONNX is designed to allow framework inter-operability. There are many excellent
+machine learning libraries in various languages : PyTorch, TensorFlow, MXNet, and Caffe are
[…]
