### 2019-11-04 d328243
modifications to the paper
