+++ paper/paper.tex
@@ -1,137 +1,0 @@
-
-% JuliaCon proceedings template
-\documentclass{juliacon}
-\setcounter{page}{1}
-
-\begin{document}
-
-\input{header}
-
-\maketitle
-
-\begin{abstract}
-
-Flux\cite{innes:2018} is a machine learning framework, written using the numerical computing language Julia\cite{bezanson2017julia}. The framework makes writing layers as simple as writing mathematical formulae, and it's advanced AD, Zygote\cite{DBLP:journals/corr/abs-1810-07951}, applies automatic differentiation (AD) to seamlessly calculate derivatives and train the model. It makes heavy use of Julia's language and compiler features to carry out code analysis and make optimisations.  For example, Julia's GPU compilation support\cite{besard:2017} can be used to JIT-compile custom GPU kernels for model layers\cite{CuArrays}. Flux also supports a number of a hardware options, from CPUs, GPUs and even including TPUs.\newline
-
-ONNX.jl is an Open Neural Network Exchange backend for the Flux.jl deep learning framework. ONNX.jl supports directly importing high quality ONNX standard models into Flux, thus saving time and reducing the need for additional computation resources.
-
-\end{abstract}
-
-\section{Introduction}
-
-Open Neural Network Exchange (ONNX) is an open ecosystem that empowers AI developers to choose the right tools as their project evolves. ONNX provides an open source format for AI models, both deep learning and traditional ML. It defines an extensible computation graph model, as well as definitions of built-in operators and standard data types. It provides a set of specifications to convert a model to a basic ONNX format, and another set of specifications to get the model back from this ONNX form.\newline  At a high level, ONNX is designed to allow framework inter-operability. There are many excellent machine learning libraries in various languages — PyTorch, TensorFlow, MXNet, and Caffe are just a few that have become very popular in recent years, but there are many others as well.
-
-\subsection{A simple use case}
-ONNX is simplifying AI choices by defining a standard which is here to stay, and is usable anywhere from small mobile devices to large server farms, across chipsets and vendors, and with extensive runtimes and tools support. ONNX reduces the friction of moving trained AI models among your favorite tools and frameworks and platforms. A simple example of how ONNX is ideal for ML is the case when large deep learning models need to be deployed. \newline \newline \newline Consider the simple case of deploying a Deep Learning model to an iOS application. This particular model can be implemented in any framework : TensorFlow, PyTorch, MXNet just to name a few. However, iOS applications expect to use CoreML inside the application. Up until now, developers have been porting large models to different frameworks, which is a waste of time and energy, better spent somewhere else. This is also retraining the entire model from scratch, which isn't efficient. This makes the entire process cumbersome and impractical. ONNX exists to solve this very problem : By connecting the common dots from different frameworks, ONNX makes it possible to express a model of type A to type B, thus saving time and the need to train the model again.
-
-\section{Structure of an ONNX serialized model:}
-ONNX uses Google's Protocol Buffers to serialize and de-serialize models. Julia has a Protocol Buffer backend via \texit{ProtoBuf.jl}. \newline
-Since ONNX tries to inherit properties from diverse frameworks, ONNX serialized models can be large and complicated. In general, the model is essentially a \textit{ModelProto} object, that consists a number of sub-fields. The \textit{ModelProto} structure looks something like:
-
-\begin{lstlisting}[language=julia]
-mutable struct ModelProto <: ProtoType
-    ir_version::Int64
-    opset_import::Vector{OperatorSetIdProto}
-    producer_name::AbstractString
-    producer_version::AbstractString
[…]
