+++ paper/header.tex
@@ -1,12 +1,0 @@
-% **************GENERATED FILE, DO NOT EDIT**************
-
-\title{Interoperating Deep Learning models with ONNX.jl}
-
-\author[1]{Ayush Shridhar}
-\author[2]{Mike Innes}
-
-\affil[1]{International Institute of Information Technology, Bhubaneswar, India}
-\affil[2]{Julia Computing}
-
-\keywords{Julia, Machine Learning, Deep Learning, Transfer Learning}
-
