### 2019-11-05 #21 @NiklasGustafsson
I'm not on the JuliaLang Slack channel. You can contact me at niklas.gustafsson@microsoft.com, so I'll have your email address. Once I have that, I will send you some thoughts.
