+++ LICENSE.md
@@ -1,6 +1,6 @@
 The ONNX.jl package is licensed under the MIT "Expat" License:
 
-> Copyright (c) 2018: Mike J Innes.
+> Copyright (c) 2018: Mike J Innes, Julia Computing & contributors.
 >
 > Permission is hereby granted, free of charge, to any person obtaining a copy
 > of this software and associated documentation files (the "Software"), to deal
