### 2020-07-27 #59 @whedon
[REVIEW]: Interoperating Deep Learning models with ONNX.jl
**Submitting author:** @ayush-1506 (<a href="http://orcid.org/0000-0003-3550-691X">Ayush Shridhar</a>)
**Repository:** <a href="https://github.com/FluxML/ONNX.jl" target ="_blank">https://github.com/FluxML/ONNX.jl</a>
**Version:**
