### 2020-07-27 #59 @vchuravy
@ayush-1506 please re-add the paper once you are done with your modifications: https://github.com/FluxML/ONNX.jl/commit/d328243656ab3264b9fd58ce6f5eb0bc3f8b1711
