+++ paper/paper.tex
@@ -1,0 +1,418 @@
+
+% JuliaCon proceedings template
+\documentclass{juliacon}
+\setcounter{page}{1}
+\usepackage{graphicx}
+
+\hypersetup{colorlinks=true}
+\begin{document}
+
+\input{header}
+
+\maketitle
+
+\begin{abstract}
+
+Flux \cite{innes:2018} is a machine learning framework, written using the numerical computing
+language Julia\cite{bezanson2017julia}. The framework makes writing layers as simple as
+writing mathematical formulae, and it's advanced AD,
+Zygote \cite{DBLP:journals/corr/abs-1810-07951} , applies automatic differentiation (AD) to
+calculate derivatives and train the model. It makes heavy use of Julia's language and
+compiler features to carry out code analysis and make optimisations.  For example, Julia's
+GPU compilation support \cite{besard:2017} can be used to JIT-compile custom GPU kernels for
+model layers \cite{CuArrays.jl}. Flux also supports a number of a hardware options, from CPUs,
+GPUs and even TPUs via XLA.jl, that compiles Julia code to XLA: an advanced compiler for
+linear algebra that is capable of greatly optimizing speed and memory usage in large
+deep learning models. 
+
+ONNX.jl is an Open Neural Network Exchange backend for the Flux.jl deep learning framework.
+ONNX.jl supports directly importing high quality ONNX standard models into Flux, thus saving
+time and reducing the need for additional computation resources. This paper aims at introducing 
+ONNX.jl and explaining how it fits into the bigger picture: How we can use the Julia Language, 
+specifically Flux.jl and ONNX.jl as a starting for high quality transfer learning of large deep
+learning models.
+
+\end{abstract}
+
+\section{Introduction}
+
+The Julia language was introduced to solve the two language problem: In simple words, languages that
+are simple to write (high-level) are very slow but those which are difficult to use (low-level) are
+way faster. This is because most of the high-level languages weren't written to process a large 
+amount of data. Thus engineers, researchers and developers have a hard time developing a lot of 
+high performance languages. At the moment, the common protocol is to write the core of the software
+in a low-level language (C/C++/Fortran) and wrap it in a high-level language (Python). This results
+in optimized performance and ease of use. The Julia language aims to make best of both worlds. It 
+provides a high level syntax but manages to perform as fast as C (sometimes even faster).  
+
+Flux.jl is a library for implementing machine learning models, written completely in the Julia
+programming language. At the heart of Flux.jl lies Zygote.jl: A source-to-source automatic
+differentiation (AD) library that makes complete use of the Julia language compiler to generate
+backward pass during training phase of a neural network, with complete support for control flow,
+recursion, closures and data structures.  
+Implementing models in Flux.jl is as simple as writing regular Julia code. Implementing models
+is as simple as writing the formulae for those, and Zygote.jl will compute the derivatives 
+seamlessly.  
+
+Flux.jl also provides support for other hardware options using external packages such as CuArrays.jl
+and CLArrays.jl. CuArrays is written completely in Julia, making implementing GPU kernels very simple.
+Making a model run on GPU can be done in a hassle-free manner: It is as simple as calling a few functions
+to transfer data to GPU. Flux.jl also has support for running models on Google's Tensor Processing
+Unit (TPU). TPUs help in very fast linear algebra computation. Running Flux models on TPUs is possible
+through XLA.jl that compiles Julia code to XLA.  
+
+The FluxML ecosystem provides a number of supporting packages that provide additional functionalities , some of them being (apart from the aforementioned Flux.jl, Zygote.jl and XLA.jl);
+
+\begin{itemize}
+    \item ONNX.jl : Open Neural Network eXchange backend for Flux.jl
[…]
