+++ paper/ref.bib
@@ -1,0 +1,257 @@
+@article{innes:2018,
+  author    = {Mike Innes},
+  title     = {Flux: Elegant Machine Learning with Julia},
+  journal   = {Journal of Open Source Software},
+  year      = {2018},
+  doi       = {10.21105/joss.00602},
+}
+
+@article{bezanson2017julia,
+  title={Julia: A fresh approach to numerical computing},
+  author={Bezanson, Jeff and Edelman, Alan and Karpinski, Stefan and Shah, Viral B},
+  journal={SIAM review},
+  volume={59},
+  number={1},
+  pages={65--98},
+  year={2017},
+  doi       = {10.1137/141000671}
+}
+
+@article{DBLP:journals/corr/abs-1810-07951,
+  author    = {Michael Innes},
+  title     = {Don't Unroll Adjoint: Differentiating SSA-Form Programs},
+  journal   = {CoRR},
+  volume    = {abs/1810.07951},
+  year      = {2018},
+  url       = {http://arxiv.org/abs/1810.07951},
+  archivePrefix = {arXiv},
+  eprint    = {1810.07951},
+  timestamp = {Tue, 30 Oct 2018 20:39:56 +0100},
+  biburl    = {https://dblp.org/rec/bib/journals/corr/abs-1810-07951},
+  bibsource = {dblp computer science bibliography, https://dblp.org}
+}
+
+@article{besard:2017,
+  author    = {Tim Besard and Christophe Foket and De Sutter, Bjorn},
+  title     = {Effective Extensible Programming: Unleashing {Julia} on {GPUs}},
+  journal   = {arXiv},
+  volume    = {abs/11712.03112},
+  year      = {2017},
+  url       = {https://arxiv.org/abs/1712.03112},
+  doi       = {10.1109/tpds.2018.2872064}
+}
+
+@online{CuArrays.jl,
+  author = {Mike Innes and others},
+  title = {Generic GPU Kernels},
+  year = 2017,
+  url = {https://mikeinnes.github.io/2017/08/24/cudanative.html},
+  urldate = {2018-02-16}
+}
+
+@online{DataFlow.jl,
+  author = {Mike Innes},
+  title = {DataFlow.jl: Code Intermediate Representation Format},
+  year = 2016,
+  url = {https://github.com/MikeInnes/DataFlow.jl},
+  urldate = {2018-09-27}
+}
+
+@online{FluxJS.jl,
+  author = {Mike Innes},
+  title = {FluxJS.jl: Flux to javascript compiler},
+  year = 2018,
+  url = {https://github.com/FluxML/FluxJS.jl},
+  urldate = {2018-09-27}
+}
+
+@online{model-zoo,
+  author = {Mike Innes},
+  title = {model-zoo: Flux models},
+  url = {https://github.com/FluxML/model-zoo},
+}
+
+@online{IRTools.jl,
+  author = {Mike Innes},
+  title = {Intermediate Representation Tool},
+  url = {https://github.com/FluxML/IRTools.jl},
+}
+
+@online{Torch.jl,
+  author = {Dhairya Gandhi},
+  title = {Sensible extensions for exposing torch in Julia. },
+  url = {https://github.com/FluxML/Torch.jl},
+}
+
+@online{Metalhead.jl,
+  author = {Mike Innes},
+  title = {Computer vision models for Flux },
+  url = {https://github.com/FluxML/Metalhead.jl},
+}
+
+@online{Protobuf.jl,
+  author = {Tanmay Mohapatra},
+  title = {Julia protobuf implementation},
+  url = {https://github.com/JuliaIO/ProtoBuf.jl},
+}
+
+@online{BSON.jl,
+  author = {Mike Innes},
[…]
