+++ paper/bib.tex
@@ -1,0 +1,4 @@
+`% **************GENERATED FILE, DO NOT EDIT**************
+
+\bibliographystyle{juliacon}
+\bibliography{ref.bib}
