+++ paper/header.tex
@@ -1,0 +1,20 @@
+% **************GENERATED FILE, DO NOT EDIT**************
+
+\title{Interoperating Deep Learning models with ONNX.jl}
+
+\author[1]{Ayush Shridhar}
+\author[2]{Phil Tomson}
+\author[3]{Mike Innes}
+
+\affil[1]{International Institute of Information Technology, Bhubaneswar, India}
+\affil[2]{Unaffiliated}
+\affil[3]{Julia Computing, Inc.}
+
+\keywords{Julia, Machine Learning, Deep Learning, Transfer Learning}
+
+\hypersetup{
+pdftitle = {Interoperating Deep Learning models with ONNX.jl},
+pdfsubject = {JuliaCon 2019 Proceedings},
+pdfauthor = {Ayush Shridhar, Phil Tomson, Mike Innes},
+pdfkeywords = {Julia, Machine Learning, Deep Learning, Transfer Learning},
+}
