### 2020-08-17 #59 @vchuravy
@ayush-1506  can you fix the missing DOIs? Also you should include references to projects/papers/work out-side the Julia ecosystem that inspired your work.
