### 2020-08-17 #59 @ekinakyurek
Hi @ayush-1506, I've two comments
-If more space allowed, I would like to see details of ModelProto and GraphProto.
-There are zigzags in Fig1.
