### 2020-08-17 #59 @ayush-1506
@ekinakyurek Thanks for the review. I've fixed these details.
